post

+++ app/Http/Controllers/TasksController.php
@@ -14,9 +14,17 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-        
-        return view('tasks.index', ['tasks'=>$tasks,]);
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+        // 認証済みユーザを取得
+        $user = \Auth::user();
+        $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+                $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+        return view('welcome', $data);
     }
 
     /**
@@ -26,9 +34,12 @@ class TasksController extends Controller
      */
     public function create()
     {
+       
         $task = new Task;
         return view('tasks.create', ['task' => $task,]);
+        
     }
+
 
     /**
      * Store a newly created resource in storage.
@@ -44,11 +55,13 @@ class TasksController extends Controller
             'content' => 'required|max:255',
         ]);
         
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-        
+        // 認証済みユーザ（閲覧者）の投稿として作成
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
+
+        // 前のURLへリダイレクトさせる
         return redirect('/');
 
     }
@@ -62,7 +75,11 @@ class TasksController extends Controller
     public function show($id)
     {
         $task = Task::findOrFail($id);
-        return view('tasks.show', ['task' => $task,]);
+        if (\Auth::id() === $task->user_id) {
+        return view('tasks.show', [
+            'task' => $task,
+        ]);}
+        else { return redirect('/'); }
     }
 
     /**
@@ -74,9 +91,10 @@ class TasksController extends Controller
     public function edit($id)
     {
         $task = Task::findOrFail($id);
+        if (\Auth::id() === $task->user_id) {
         return view('tasks.edit', ['task' => $task,]);
-    }
-
+        }else { return redirect('/'); }}
+        
     /**
      * Update the specified resource in storage.
      *
@@ -93,9 +111,11 @@ class TasksController extends Controller
         ]);
         
         $task = Task::findOrFail($id);
+        if (\Auth::id() === $task->user_id) {
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
+        }
         
         return redirect('/');
     }
@@ -108,9 +128,16 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::findOrFail($id);
-        $task->delete();
-        
+        // idの値で投稿を検索して取得
+        $task = \App\Task::findOrFail($id);
+
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
+        // 前のURLへリダイレクトさせる
         return redirect('/');
     }
 }
+
